Capstone II Project - Richard/requirements.txt
Flask==3.1.3
flask_cors==6.0.2
numpy==2.4.4
Pillow==12.2.0
pygltflib==1.16.5
trimesh==4.11.2
pybind11==3.0.2
